requirements.txt
pandas
beautifulsoup4
pandas-gbq
python-dotenv